Add -l list option to only list filenames
The `sunzip -l` option is equivalent to `-t -q -q`, but instead
of testing decompression will only print the file and directory
names as they are encountered in the stream. Note that file
names from the local file headers are less reliable than the
end-of-file TOC that would otherwise be used, and may
include duplicates, deleted and encrypted files.