Add a Makefile and the source files for Deflate64 and PKWare DCL.
This commit adds the source code for decompressing Deflate64 and
PKWare DCL zip file entries, copied from zlib's contrib directory.
A Makefile is added to compile those and to link libbzip2 as well
for bzip2 decompression. Compile-time options are added to include
the compression methods, DEFLATE64, PKDCL, and BZIP2. Any or all of
those can be removed in the Makefile compile command. With none of
them, sunzip can decompress stored and deflated entries, which
covers the mast majority of zip files found in the wild.